zmiany poprawki zabezpieczenia

+++ restore_nots.php
@@ -1,19 +1,24 @@
-<?php session_start();
-
-require_once "database/dbinfo.php";
+<?php
+session_start();
+if (!empty($_POST)){
+    require_once "database/dbinfo.php";
     require_once "objects.php";
     $connection = db_connection();
-     if ($connection != false){ 
-         
-         if(isset($_POST['delete'])){
-             foreach($_POST['not'] as $selected){
-                 $sql="DELETE  FROM $db_nots_user_tab WHERE  $db_nots_user_id = $selected;";
-                $connection->query($sql); }   
-         }else{
-          
+    if ($connection != false){ 
+        if(isset($_POST['delete'])){
             foreach($_POST['not'] as $selected){
-                 $sql="UPDATE $db_nots_user_tab SET $db_nots_user_delete = '0' WHERE $db_nots_user_id = $selected;";
-                 $connection->query($sql);
-         } }         
+                $sql="DELETE  FROM $db_nots_user_tab WHERE  $db_nots_user_id = $selected;";
+                $connection->query($sql); 
+            }   
+        }
+        else{
+            foreach($_POST['not'] as $selected){
+                $sql="UPDATE $db_nots_user_tab SET $db_nots_user_delete = '0' WHERE $db_nots_user_id = $selected;";
+                $connection->query($sql);
+            }
+        }
+        $connection->close();
     }
-        header("location: bin.php");
+}
+header("location: bin.php");
+?>
